move price formatting to utils
