+++ src/components/ui/Card.jsx
@@ -6,6 +6,7 @@ import DoughnutChart from './DoughnutChart';
 import { NutriButton } from './Buttons';
 import NutritionalValues from './NutritionalValuesContainer';
 import { Nutriscore } from './Nutriscore';
+import { formatPrice } from '../../utils/price';
 
 /** Custom wrapper for cards with images
  * @param {object} styles styles passed from the global stylesheet
@@ -43,6 +44,7 @@ const MealCard = (
     const {t} = useTranslation();
     const [expanded, setExpanded] = useState(false);
 
+    const formattedPrice = formatPrice(meal.price);
     const nutrition = {
         energy: meal.energy,
         carbohydrates: meal.carbohydrates,
@@ -112,20 +114,6 @@ const MealCard = (
             </View>
         );
     };
-
-    const formatPrice = (price) => {
-        let formatted =(price/100).toString().replace(/\./, ',');
-        if (formatted.indexOf(',') !== -1) {
-            const fractionalPart = formatted.split(',')[1];
-            if (fractionalPart.length === 1) {
-                formatted += '0';
-            }
-        } else {
-            formatted += ',00';
-        }
-        return formatted += ' €';
-    };
-    const formattedPrice = formatPrice(meal.price);
 
     return (
         <View style={styles.cardContainer}>
